refactor: 영화 후기 포스팅, 포스팅 댓글 인터페이스 분리
